% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareBnetPopGenoPDFs.R
\name{compareBnetPopGenoPDFs}
\alias{compareBnetPopGenoPDFs}
\title{Compare population and Bayesian network genotype probability density functions # nolint}
\usage{
compareBnetPopGenoPDFs(lprobTable)
}
\arguments{
\item{lprobTable}{list of probability tables}
}
\value{
list of KL divergences
}
\description{
Compare population and Bayesian network genotype probability density functions # nolint
}
