% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPed.R
\name{convertPed}
\alias{convertPed}
\title{Convert a pedigree to a paramlink object}
\usage{
convertPed(x, verbose = FALSE)
}
\arguments{
\item{x}{pedigree}

\item{verbose}{print progress}
}
\value{
paramlink object
}
\description{
Convert a pedigree to a paramlink object
}
