% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossH.R
\name{crossH}
\alias{crossH}
\title{Cross entropy}
\usage{
crossH(px, py, epsilon = 1e-20)
}
\arguments{
\item{px}{probability distribution}

\item{py}{probability distribution}

\item{epsilon}{small number to avoid log(0)}
}
\value{
cross entropy
}
\description{
Cross entropy
}
