% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elimLangeGoradia.R
\name{elimLangeGoradia}
\alias{elimLangeGoradia}
\title{Eliminate Mendelian errors using Lange-Goradia algorithm}
\usage{
elimLangeGoradia(ped, iMarker = 1, bitera = TRUE, bverbose = TRUE)
}
\arguments{
\item{ped}{pedigree}

\item{iMarker}{index of marker to be used}

\item{bitera}{iterate until no more errors are found}

\item{bverbose}{print progress}
}
\value{
pedigree with Mendelian errors eliminated
}
\description{
Eliminate Mendelian errors using Lange-Goradia algorithm
}
