% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPed.R
\name{exportPed}
\alias{exportPed}
\title{Export a pedigree to a file}
\usage{
exportPed(ped, fname, iMarker = 1)
}
\arguments{
\item{ped}{pedigree}

\item{fname}{file name}

\item{iMarker}{index of marker to be used}
}
\value{
pedigree with Mendelian errors eliminated
}
\description{
Export a pedigree to a file
}
