% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensIT.R
\docType{package}
\name{forensIT}
\alias{forensIT-package}
\alias{forensIT}
\title{forensIT: Information Theory Tools for Forensic Analysis}
\description{
The 'forensIT' package, available on CRAN, is a comprehensive statistical toolkit  tailored for handling missing person cases. By leveraging information theory metrics, it enables accurate assessment of kinship, particularly when limited genetic evidence is available.   With a focus on optimizing statistical power, 'forensIT' empowers investigators to effectively prioritize family members, enhancing the reliability and efficiency of missing person investigations. Experience the power of information theory in kinship testing with the user-friendly 'forensIT' package, freely accessible on CRAN. # nolint
}
\author{
\strong{Maintainer}: Franco Marsico \email{franco.lmarsico@gmail.com}

Authors:
\itemize{
  \item Ariel Chernomoretz \email{ariel@df.uba.ar}
}

}
