% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeProbs.R
\name{genotypeProbs}
\alias{genotypeProbs}
\title{Genotype probabilities}
\usage{
genotypeProbs(probP, probM)
}
\arguments{
\item{probP}{vector of parental probabilities}

\item{probM}{vector of parental probabilities}
}
\value{
matrix of genotype probabilities
}
\description{
Calculate genotype probabilities from parental probabilities
}
