% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeProbTable.R
\name{genotypeProbTable}
\alias{genotypeProbTable}
\title{Genotype Probability Table}
\usage{
genotypeProbTable(bbn1, resQQ, bplot = FALSE, numMarkers = 4, lLoci)
}
\arguments{
\item{bbn1}{Bayesian network}

\item{resQQ}{results from bn}

\item{bplot}{boolean to plot}

\item{numMarkers}{number of markers}

\item{lLoci}{list of loci}
}
\value{
Genotype Probability Table
}
\description{
Genotype Probability Table
}
