% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeProbTable_FM.R
\name{genotypeProbTable_FM}
\alias{genotypeProbTable_FM}
\title{genotypeProbTable_FM}
\usage{
genotypeProbTable_FM(bbn1, resQQ, bplot = FALSE, numMarkers = 4, freq)
}
\arguments{
\item{bbn1}{bayesian network}

\item{resQQ}{list of results from the inference}

\item{bplot}{plot results}

\item{numMarkers}{number of markers}

\item{freq}{allele frequencies}
}
\value{
matrix of genotype probabilities
}
\description{
function to calculate the probability of genotypes given the MOI
}
