% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllelesFromGenotypes.R
\name{getAllelesFromGenotypes}
\alias{getAllelesFromGenotypes}
\title{getAllelesFromGenotypes}
\usage{
getAllelesFromGenotypes(g)
}
\arguments{
\item{g}{genotypes}
}
\value{
alleles
}
\description{
Get alleles from genotypes
}
