% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{Entropy of a discrete probability distribution}
\usage{
H(px, epsilon = 1e-20, normalized = FALSE)
}
\arguments{
\item{px}{probability distribution}

\item{epsilon}{small number to avoid log(0)}

\item{normalized}{boolean to normalize entropy}
}
\value{
entropy
}
\description{
Entropy of a discrete probability distribution
}
