% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2Genotypes.pedtools.R
\name{index2Genotypes2.pedtools}
\alias{index2Genotypes2.pedtools}
\title{index2Genotypes}
\usage{
index2Genotypes2.pedtools(ped, id, iMarker, alleleSet)
}
\arguments{
\item{ped}{pedigree}

\item{id}{individual id}

\item{iMarker}{marker index}

\item{alleleSet}{allele set}
}
\value{
genotypes
}
\description{
index2Genotypes
}
