% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidateId.R
\name{InvalidateId}
\alias{InvalidateId}
\title{invalidateId}
\usage{
InvalidateId(ped, id = NULL, markerId = 1, double = FALSE, propagate = FALSE)
}
\arguments{
\item{ped}{pedigree}

\item{id}{individual's id}

\item{markerId}{marker's id}

\item{double}{if TRUE, invalidate both alleles}

\item{propagate}{if TRUE, propagate the invalidation to descendants}
}
\value{
pedigree with invalidated individual
}
\description{
Invalidate an individual
}
