% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLd.R
\name{KLd}
\alias{KLd}
\title{KL divergence}
\usage{
KLd(ppx, ppy, epsilon = 1e-20, bsigma = FALSE)
}
\arguments{
\item{ppx}{probability distribution}

\item{ppy}{probability distribution}

\item{epsilon}{small number to avoid log(0)}

\item{bsigma}{boolean to compute sigma}
}
\value{
KL divergence
}
\description{
KL divergence
}
