% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLde.R
\name{KLde}
\alias{KLde}
\title{KL divergence}
\usage{
KLde(px, py, epsilon = 1e-20)
}
\arguments{
\item{px}{probability distribution}

\item{py}{probability distribution}

\item{epsilon}{small number to avoid log(0)}
}
\value{
KL divergence
}
\description{
KL divergence
}
