% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perMarkerKLs.R
\name{perMarkerKLs}
\alias{perMarkerKLs}
\title{perMarkerKLs}
\usage{
perMarkerKLs(ped, MP, frequency)
}
\arguments{
\item{ped}{Reference pedigree.}

\item{MP}{missing person}

\item{frequency}{Allele frequency database.}
}
\value{
An object of class data.frame with KLs.
}
\description{
perMarkerKLs
}
\examples{
library(forrel)
x = linearPed(2)
plot(x)
x = setMarkers(x, locusAttributes = NorwegianFrequencies[1:5])
x = profileSim(x, N = 1, ids = 2)
perMarkerKLs(x, MP = 5 , NorwegianFrequencies[1:5])
}
