% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKL.R
\name{plotKL}
\alias{plotKL}
\title{plotKL: scatter plot of obtained KL values from distKL}
\usage{
plotKL(res)
}
\arguments{
\item{res}{output from distKL function.}
}
\value{
A scatterplot.
}
\description{
plotKL: scatter plot of obtained KL values from distKL
}
\examples{
library(forrel)
x = linearPed(2)
plot(x)
x = setMarkers(x, locusAttributes = NorwegianFrequencies[1:5])
x = profileSim(x, N = 1, ids = 2)
res <- distKL(ped = x, missing = 5, relative = 1, cores = 10, frequency = NorwegianFrequencies[1:5], numsims = 5)
plotKL(res)
}
