% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Px.R
\name{Px}
\alias{Px}
\title{Px}
\usage{
Px(p1, p0, dbg = FALSE)
}
\arguments{
\item{p1}{probability distribution}

\item{p0}{probability distribution}

\item{dbg}{boolean to compute sigma}
}
\value{
Px
}
\description{
Px
}
