% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-it.R
\name{runIT}
\alias{runIT}
\title{runIT}
\usage{
runIT(
  lped = NULL,
  freqs,
  QP,
  dbg,
  numCores,
  bOnlyIT = FALSE,
  lprobg_ped = NULL,
  bsigma = FALSE,
  blog = FALSE,
  dep = TRUE
)
}
\arguments{
\item{lped}{list of pedigree objects}

\item{freqs}{list of allele frequencies}

\item{QP}{QP}

\item{dbg}{debug}

\item{numCores}{number of cores}

\item{bOnlyIT}{boolean to only run IT}

\item{lprobg_ped}{list of probG}

\item{bsigma}{boolean to compute sigma}

\item{blog}{boolean to write log}

\item{dep}{check fbnet dependency}
}
\value{
runIT
}
\description{
run information theory (IT) metrics
}
