% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLR.R
\name{simLR}
\alias{simLR}
\title{Simulate LR}
\usage{
simLR(
  lprobg_ped,
  numSim = 10000,
  epsilon = 1e-20,
  bplot = FALSE,
  bLRs = FALSE,
  seed = 123457
)
}
\arguments{
\item{lprobg_ped}{list of probability distributions}

\item{numSim}{number of simulations}

\item{epsilon}{small number to avoid log(0)}

\item{bplot}{boolean to plot}

\item{bLRs}{boolean to return LRs}

\item{seed}{seed}
}
\value{
LRs
}
\description{
Simulate LR
}
