% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simME.R
\docType{data}
\name{simME}
\alias{simME}
\title{simME: output from simMinimalEnsemble considering an uncle}
\format{
A list with minimalEnsemble of genotypes
}
\usage{
simME
}
\description{
simME: output from simMinimalEnsemble considering an uncle
}
\keyword{datasets}
