% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTestIDMarkers.R
\name{simTestIDMarkers}
\alias{simTestIDMarkers}
\title{Simulate testID markers}
\usage{
simTestIDMarkers(ped, testID, numSim = 10, seed = 123457)
}
\arguments{
\item{ped}{pedigree}

\item{testID}{test ID}

\item{numSim}{number of simulations}

\item{seed}{seed}
}
\value{
list of simulations
}
\description{
Simulate testID markers
}
