% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trioCheckFast.R
\name{trioCheckFast}
\alias{trioCheckFast}
\title{trioCheckFast}
\usage{
trioCheckFast(ffa, mmo, oof)
}
\arguments{
\item{ffa}{father's alleles}

\item{mmo}{mother's alleles}

\item{oof}{offspring's alleles}
}
\value{
TRUE if there is a Mendelian error
}
\description{
Check for Mendelian errors in trios
}
