% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidimKLplot.R
\name{unidimKLplot}
\alias{unidimKLplot}
\title{unidimKLplot: KL distributions presented in the same units (Log10(LR))}
\usage{
unidimKLplot(res)
}
\arguments{
\item{res}{output from distKL function.}
}
\value{
A scatterplot.
}
\description{
unidimKLplot: KL distributions presented in the same units (Log10(LR))
}
