% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnsembleCPTs.R
\name{buildEnsembleCPTs}
\alias{buildEnsembleCPTs}
\title{buildEnsembleCPTs}
\usage{
buildEnsembleCPTs(lsimu, lminimalProbGenoMOI)
}
\arguments{
\item{lsimu}{list of simulations}

\item{lminimalProbGenoMOI}{list of minimal probabilities of genotypes given MOI # nolint}
}
\value{
list of CPTs
}
\description{
Build ensemble of CPTs from a list of simulations
}
