% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnsembleITValues.R
\name{buildEnsembleITValues}
\alias{buildEnsembleITValues}
\title{buildEnsembleITValues}
\usage{
buildEnsembleITValues(
  lsimu = lsimulation,
  ITtab = sim$ITtable,
  bFullIT = FALSE
)
}
\arguments{
\item{lsimu}{list of simulations}

\item{ITtab}{IT table}

\item{bFullIT}{boolean to return full IT table}
}
\value{
list of IT values
}
\description{
Build ensemble of IT values from a list of simulations
}
