% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPed.R
\name{convertPed}
\alias{convertPed}
\title{Convert a pedigree to a paramlink object}
\usage{
convertPed(x, verbose = FALSE)
}
\arguments{
\item{x}{pedigree}

\item{verbose}{print progress}
}
\value{
paramlink object
}
\description{
Convert a pedigree to a paramlink object
}
\examples{
library(forrel)
x = linearPed(2)
plot(x)
x = setMarkers(x, locusAttributes = NorwegianFrequencies[1:2])
x = profileSim(x, N = 1, ids = 2)
convertPed(x)
}
