% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perMarkerKLs.R
\name{perMarkerKLs}
\alias{perMarkerKLs}
\title{perMarkerKLs: Obtain KL per maker for a specific pedigree}
\usage{
perMarkerKLs(ped, frequency)
}
\arguments{
\item{ped}{Reference pedigree. It could be an input from read_fam() function or a pedigree built with pedtools.}

\item{frequency}{Allele frequency database.}
}
\value{
An object of class data.frame with KLs.
}
\description{
perMarkerKLs: Obtain KL per maker for a specific pedigree
}
\examples{
library(forrel)
x = linearPed(2)
plot(x)
x = setMarkers(x, locusAttributes = NorwegianFrequencies[1:5])
x = profileSim(x, N = 1, ids = 2)
perMarkerKLs(x,  NorwegianFrequencies[1:5])
}
