% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMinimalEnsemble.R
\name{simMinimalEnsemble}
\alias{simMinimalEnsemble}
\title{simMinimalEnsemble}
\usage{
simMinimalEnsemble(
  ped,
  QP,
  testID,
  freqs,
  numCores = 1,
  seed = 123457,
  bVerbose = TRUE,
  bJustGetNumber = FALSE,
  bdbg = FALSE
)
}
\arguments{
\item{ped}{pedigree}

\item{QP}{QP}

\item{testID}{test ID}

\item{freqs}{frequencies}

\item{numCores}{number of cores}

\item{seed}{seed}

\item{bVerbose}{boolean to print information}

\item{bJustGetNumber}{boolean to just get the number of runs}

\item{bdbg}{boolean to debug}
}
\value{
list of results
}
\description{
simMinimalEnsemble
}
